Set wait time to all the client requests
